requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
pydantic-settings
gradio
PyMuPDF
pypdf
pdfplumber
reportlab
tenacity
pytest
